Fix selected displays not updating
When updating Placeholder distribution, switching
to HTML view, and saving, the Distribution displays
are lost

updateDistribution takes Distribution from the Placeholder list and
updates the presentation.distribution. However, when in
HTML mode, the Placeholder list is refreshed (cleared) based
on the HTML, so the Distribution is lost. This can be corrected
via parseDistribution, but that call is not needed
in this case because it's called after the update.

Also, fixing unit tests. Promise results do not count as failures
since they are not linked to the done sent in the timeout. Moving
timeout to Promise results.

[stage-2]