Add an encoding bypass session storage item
Previously e2e tests were not checking the live master switch. They were
checking a fake, so that it would always return false unless the
force-encoding session storage flag was set.

With this change, the e2e tests will use the live master switch, but in
most cases they will ignore the master switch setting because they've
set the bypass session flag. One e2e test will run without bypass so
that it uses the encoding.

The advantage is that if the live master switch is off because the
encoding service is down, the e2e test will respect the switch setting
and upload directly to GCS rather than failing. In most cases, when the
live master switch is on, the E2E test will use the encoding service and
confirm functionality.

[stage-6]